mon: split osdmon crush (part 1)
